chore: tranform form values to api values
